+++ src/vector/arithmetic/vector_mul_d.cpp
@@ -8,7 +8,7 @@ using namespace ddavx_core;
 
 void d_real_vector::mul(dd_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
+	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -16,7 +16,7 @@ void d_real_vector::mul(dd_real_vector& vec1, dd_real_vector& vec2)
 
 #pragma omp parallel private(regs)
 	{
-		long i=0, is=0, ie=0;
+		size_t i=0, is=0, ie=0;
 		get_isie(size(), is, ie);
 		for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
 			AVXreg a_hi = load(data()[i]);
@@ -39,7 +39,7 @@ void d_real_vector::mul(dd_real_vector& vec1, dd_real_vector& vec2)
 
 void d_real_vector::mul(dd_real_vector& vec1, d_real_vector& vec2)
 {
-	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
+	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -48,7 +48,7 @@ void d_real_vector::mul(dd_real_vector& vec1, d_real_vector& vec2)
 
 #pragma omp parallel private(regs)
 	{
-		long i=0, is=0, ie=0;
+		size_t i=0, is=0, ie=0;
 		get_isie(size(), is, ie);
 		for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
 			AVXreg a_hi = load(data()[i]);
@@ -76,13 +76,13 @@ void d_real_vector::mul(d_real_vector& vec1, dd_real_vector& vec2)
 
 void d_real_vector::mul(d_real_vector& vec1, d_real_vector& vec2)
 {
-	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
+	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
 
 #pragma omp parallel for
-	for(long i = 0; i < (long)size(); i++){
+	for(size_t i = 0; i < (size_t)size(); i++){
 		data()[i] = vec1.data()[i] * vec2.data()[i];
 
 	}
