+++ src/vector/arithmetic/vector_mul_dd.cpp
@@ -8,7 +8,7 @@ using namespace ddavx_core;
 
 void dd_real_vector::mul(dd_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
+	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -16,7 +16,7 @@ void dd_real_vector::mul(dd_real_vector& vec1, dd_real_vector& vec2)
 
 #pragma omp parallel private(regs)
 	{
-		long i=0, is=0, ie=0;
+		size_t i=0, is=0, ie=0;
 		get_isie(size(), is, ie);
 		for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
 			AVXreg a_hi = load(hi[i]);
@@ -41,7 +41,7 @@ void dd_real_vector::mul(dd_real_vector& vec1, dd_real_vector& vec2)
 
 void dd_real_vector::mul(dd_real_vector& vec1, d_real_vector& vec2)
 {
-	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
+	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -49,7 +49,7 @@ void dd_real_vector::mul(dd_real_vector& vec1, d_real_vector& vec2)
 
 #pragma omp parallel private(regs)
 	{
-		long i=0, is=0, ie=0;
+		size_t i=0, is=0, ie=0;
 		get_isie(size(), is, ie);
 		for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
 			AVXreg a_hi = load(hi[i]);
@@ -79,7 +79,7 @@ void dd_real_vector::mul(d_real_vector& vec1, dd_real_vector& vec2)
 
 void dd_real_vector::mul(d_real_vector& vec1, d_real_vector& vec2)
 {
-	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
+	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -87,7 +87,7 @@ void dd_real_vector::mul(d_real_vector& vec1, d_real_vector& vec2)
 
 #pragma omp parallel private(regs)
 	{
-		long i=0, is=0, ie=0;
+		size_t i=0, is=0, ie=0;
 		get_isie(size(), is, ie);
 		for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
 			AVXreg a_hi = load(hi[i]);
